Store tokens as decoded byte arrays
This avoids spending ~50-100ms converting from base32 to the
bytes used for the TOTP HMAC-SHA1 every time a code is generated.

Existing stored tokens will be converted when the app is first
launched, and new tokens will be converted when they are
added in settings.
