+++ common/util.js
@@ -57,3 +57,42 @@ function monoDigit(digit) {
     default: return digit;
   }
 }
+
+let base32chars = "ABCDEFGHIJKLMNOPQRSTUVWXYZ234567";
+
+// Convert a base32-encoded string to a Uint8Array
+export function base32ToUint8Array(base32) {
+  base32 = base32.replace(/ +/g, ""); // Strip whitespace
+
+  let array = new Uint8Array(Math.floor(base32.length * 5 / 8));
+  let arrayIndex = 0;
+  let bits = 0;
+  let value = 0;
+
+  for (let i=0; i<base32.length; i++) {
+    let char = base32.charAt(i).toUpperCase();
+    if (char === '0') {
+        char = 'O';
+    } else if (char === '8') {
+        char = 'B';
+    } else if (char === '1') {
+        char = 'L';
+    }
+    let val = base32chars.indexOf(char);
+    if ( val >= 0 && val < 32 ) {
+      value = (value << 5) | val;
+      bits += 5;
+
+      // Transfer a byte into the Uint8Array if there is enough data
+      if (bits >= 8) {
+        array[arrayIndex++] = (value >>> (bits - 8)) & 0xFF;
+        value = value & 0xFF;
+        bits -= 8;
+      }
+    } else {
+      throw Error("Character out of range: " + char);
+    }
+  }
+
+  return array;
+}
